fix: fix timestamp bugs

+++ project/stm32f407/driver/src/stm32f407_driver_hcsr04_interface.c
@@ -45,7 +45,7 @@
 /**
  * @brief timer var definition
  */
-static uint32_t gs_ms = 0;        /**< ms */
+static uint64_t gs_ms = 0;        /**< ms */
 
 /**
  * @brief     timer callback
